added stuff

+++ node2/game.h
@@ -2,7 +2,7 @@
 #define GAME_H
 
 void game_init(void);
-int game_over(void);
+uint8_t game_over(void);
 void game_play(void);
 
 #endif //GAME_H
